Flush all the buffers if the state watermark is super far behind
In the event that multiple streams are emitting data but at very different rates, one slowly filling buffer can lag the safe emission watermark to far behind what's been flushed in the other buffers. In the presence of queued STATE messages this is undesirable because they end up being emitted far later than they need to be (often at the end of the run). That means the whole run has to go perfectly well before any state messages get emitted in this situation.